Change cell_saver_ml type to int, handle large ml

+++ backend/api/migrations/0002_manual_view.py
@@ -137,7 +137,7 @@ class Migration(migrations.Migration):
                 plt_units SMALLINT UNSIGNED DEFAULT 0,
                 cryo_units SMALLINT UNSIGNED DEFAULT 0,
                 whole_units SMALLINT UNSIGNED DEFAULT 0,
-                cell_saver_ml SMALLINT UNSIGNED DEFAULT 0,
+                cell_saver_ml INT UNSIGNED DEFAULT 0,
                 overall_units SMALLINT UNSIGNED AS (rbc_units + ffp_units + plt_units + cryo_units + whole_units) STORED,
 
                 los DECIMAL(6,2),
